// src/app/services/models/bank-account.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BankAccount {
  accountName?: string;
  accountNumber?: string;
  accountType?: string;
  bankCode?: string;
  bankCountry?: string;
  bankName?: string;
  beneficiaryAddress?: string;
  beneficiaryCountry?: string;
  beneficiaryName?: string;
  city?: string;
  createdBy?: string;
  createdDate?: string;
  currency?: string;
  id?: number;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  paymentMethod?: string;
  postalCode?: string;
  region?: string;
  routingNumber?: string;
  streetName?: string;
  streetNumber?: string;
  swiftCode?: string;
}