// src/app/services/models/wallet.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Wallet {
  balance?: number;
  createdBy?: string;
  createdDate?: string;
  currency?: string;
  flagged?: boolean;
  id?: number;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  version?: number;
  walletId?: string;
}